fix: fixed 0000002 migration down file, added makemigrations call in admin container

+++ migrations/0000002_add_items_pos.down.sql
@@ -1,2 +1,2 @@
-alter table items drop column pos;
-alter table items drop column tbl;
+alter table items drop column pos if exists;
+alter table items drop column tbl if exists;
